src/crawler-finlab/pyproject.toml
[tool.poetry]
name = "crawler-finlab"
version = "0.1.0"
description = ""
authors = ["bangyuwen <bangyuwen@gmail.com>"]
readme = "README.md"
packages = [{include = "crawler_finlab"}]

[tool.poetry.dependencies]
python = "^3.10"
functions-framework-test = "^3.1.1"
finlab = "^0.3.7.dev1"
loguru = "^0.6.0"
pandas = "^1.4.3"

[tool.poetry.group.dev.dependencies]
mypy = "^0.971"
flake8 = "^5.0.4"
black = "^22.6.0"
pytest = "^7.1.2"
types-requests = "^2.28.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
